refactor(clients/java): introduce interfaces for testing
Introduce `JobPoller` and `JobRunnableFactory` interfaces (extracted
from existing implementation classes) to allow unit testing for the
metrics. This avoids the need of complex, asynchronous tests when we
just want to ensure metrics are properly called.

Some existing tests may benefit from being refactored with controllable
test implementations in the future as well.
